Make BFS use doubles for distances

+++ src/main/kotlin/graphClasses/BFS.kt
@@ -5,18 +5,18 @@ class BFS (val graph: AdjacencyList) {
 
     val size = graph.size
     var visited = BooleanArray(size)
-    val distances = IntArray(size)
+    val distances = DoubleArray(size)
     var currentVisited = mutableListOf<Int>()
-    var currentVisitedDistances = mutableListOf<Int>()
+    var currentVisitedDistances = mutableListOf<Double>()
     val parents = IntArray(graph.size) { -1 }
 
     fun bfsIterative(startIds: List<Int>) {
         currentVisited.clear()
-        distances.fill(-1)
+        distances.fill(-1.0)
         val queue = ArrayDeque<Int>()
         startIds.forEach {
             queue.add(it)
-            distances[it] = 0
+            distances[it] = 0.0
         }
         while (queue.isNotEmpty()) {
             val currentId = queue.first()
@@ -31,7 +31,7 @@ class BFS (val graph: AdjacencyList) {
                 if(!visited[v]) {
                     queue.add(v)
                     parents[v] = currentId
-                    distances[v] = (currentDistance + d).toInt()
+                    distances[v] = (currentDistance + d)
                 }
             }
         }
@@ -40,9 +40,9 @@ class BFS (val graph: AdjacencyList) {
 
     fun bfsRecursive(startIds: List<Int>) {
         currentVisited.clear()
-        distances.fill(-1)
+        distances.fill(-1.0)
         startIds.forEach {
-            distances[it] = 0
+            distances[it] = 0.0
         }
         DeepRecursiveFunction<ArrayDeque<Int>, Unit> { queue ->
             if(queue.isEmpty())
@@ -54,7 +54,7 @@ class BFS (val graph: AdjacencyList) {
             graph[id].forEach { (d, v) ->
                 if(!visited[v]) {
                     queue.add(v)
-                    distances[v] = (distances[id] + d).toInt()
+                    distances[v] = (distances[id] + d)
                 }
             }
             this.callRecursive(queue)
